refactor: move isChildNode() to utils.ts
